+++ include/utils.sh
@@ -39,6 +39,14 @@ function copy_config() {
   cp -R ${afc_path}/config/macros "${afc_config_dir}/"
 }
 
+function copy_openams_config() {
+  mkdir -p "${afc_config_dir}"
+  cp ${afc_path}/config/AFC.cfg "${afc_config_dir}/"
+  cp ${afc_path}/config/AFC_Macro_Vars.cfg "${afc_config_dir}/"
+  mkdir -p "${afc_config_dir}/mcu"
+  cp -R ${afc_path}/config/macros "${afc_config_dir}/"
+}
+
 get_git_version() {
   cd "$afc_path"
 	git_hash=$(git -C . rev-parse --short HEAD)
@@ -150,7 +158,10 @@ restart_klipper() {
 
 exit_afc_install() {
   if [ "$files_updated_or_installed" == "True" ]; then
-    restart_klipper
+    if [ "$test_mode" != "True" ]; then
+      print_msg INFO "Restarting Klipper service to apply changes..."
+      restart_klipper
+    fi
   fi
   remove_vars_tool_file
   exit 0